fix: store config/pipeline hashes as int64 to fit the uint32 range (#233)
The pipeline and Vector config hashes are CRC32 checksums (uint32,
0..4294967295), but their status fields were int32 (*uint32 in Go makes
controller-gen emit format: int32). Any hash above the int32 ceiling
(2147483647) - roughly half of all values - is rejected by the API server:

    status: Invalid value: "": Checked value must be of type integer
    with format int32 in LastAppliedPipelineHash

The status update then fails on every reconcile, so the resource stays
stuck with configCheckResult=false and the reconcile loop retries forever.
VectorPipeline only appeared unaffected when its hash happened to land
below the int32 ceiling; the defect is not specific to ClusterVectorPipeline.

Store the hashes as int64, which fully contains the uint32 range:
  - VectorPipelineStatus.LastAppliedPipelineHash
  - VectorCommonStatus.LastAppliedConfigHash
  - VectorCommonStatus.LastAppliedGlobalConfigHash

int64 is backward compatible with the int32-range values already persisted
in status. CRDs regenerated with controller-gen v0.16.1 and the helm chart
CRD copies synced.

Fixes #232

+++ api/v1alpha1/vector_common_types.go
@@ -8,10 +8,13 @@ import (
 )
 
 type VectorCommonStatus struct {
-	ConfigCheckResult           *bool   `json:"configCheckResult,omitempty"`
-	Reason                      *string `json:"reason,omitempty"`
-	LastAppliedConfigHash       *uint32 `json:"LastAppliedConfigHash,omitempty"`
-	LastAppliedGlobalConfigHash *uint32 `json:"LastAppliedGlobalConfigHash,omitempty"`
+	ConfigCheckResult *bool   `json:"configCheckResult,omitempty"`
+	Reason            *string `json:"reason,omitempty"`
+	// Config hashes are CRC32 (uint32) values stored as int64: a uint32 can exceed the
+	// int32 upper bound (2147483647), and an int32 status field rejects such values,
+	// wedging the reconcile loop with configCheckResult=false. See #232.
+	LastAppliedConfigHash       *int64 `json:"LastAppliedConfigHash,omitempty"`
+	LastAppliedGlobalConfigHash *int64 `json:"LastAppliedGlobalConfigHash,omitempty"`
 }
 
 type VectorCommon struct {
